+++ whitson/client/_whitson_client.py
@@ -1,5 +1,12 @@
-import requests
+from typing import Any, Dict, Optional
 
+from requests import Response
+
+from whitson.client._api_client import APIClient
+from whitson.client.api.fields import FieldsAPI
+from whitson.client.api.production_data import ProductionDataAPI
+from whitson.client.api.projects import ProjectsAPI
+from whitson.client.api.wells import WellsAPI
 from whitson.client.config import ClientConfig
 
 
@@ -15,63 +22,18 @@ class WhitsonClient:
     def __init__(self, config: ClientConfig):
         if config is None:
             raise ValueError("No ClientConfig has been provided.")
-        else:
-            self.config = config
-            self.base_url = f"http://{config.client_name}.whitson.com/api-external/v1/"
 
-    def get(self, suffix: str, params: dict) -> dict:
-        """Perform a GET request to an arbitrary path in the API.
+        self.config = config
+        self._api_client = APIClient(self.config)
+        self.fields = FieldsAPI(self.config)
+        self.wells = WellsAPI(self.config)
+        self.projects = ProjectsAPI(self.config)
+        self.production_data = ProductionDataAPI(self.config)
 
-        Parameters
-        ----------
-        suffix : str
-            Suffix to add to the base url for API request
-        params : dict
-            Dictionary of parameters for the request
+    def get(self, url: str, params: Optional[Dict[str, Any]] = None) -> Response:
+        """Perform a generic GET request to an arbitrary path in the API."""
+        return self._api_client.get(url=url, params=params)
 
-        Returns
-        -------
-        res : dict
-        """
-        response = requests.get(
-            url=self.base_url + suffix,
-            headers={
-                "content_type": "application/json",
-                "Authorization": f"Bearer {self.config.token.access_token}",
-            },
-            params=params,
-        )
-        res = response.json()
-        if not res:
-            raise Exception("No information retrieved")
-        return res
-
-    def post(self, suffix: str, payload: dict):
-        """Perform a POST request to an arbitrary path in the API.
-
-        Parameters
-        ----------
-        suffix : str
-            Suffix to add to the base url for API request
-        payload : dict
-            Dictionary of parameters for the request
-
-        Returns
-        -------
-        res : dict
-        """
-        response = requests.post(
-            url=self.base_url + suffix,
-            headers={
-                "content_type": "application/json",
-                "Authorization": f"Bearer {self.config.token.access_token}",
-            },
-            json=payload,
-        )
-        if 200 <= response.status_code < 300:
-            print("Successfully uploaded!")
-        else:
-            return response.text
-
-    def put(self):
-        pass
+    def post(self, url: str, payload: Optional[Dict[str, Any]] = None) -> str:
+        """Perform a generic POST request to an arbitrary path in the API."""
+        return self._api_client.post(url=url, payload=payload)