+++ CHANGELOG.md
@@ -13,6 +13,20 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.2.0] - 2023-08-16
+### Added
+- `APIClient` to generalize calls to API. Include generic get and post methods as well as some static methods to
+  retrieve metadata
+- Remove 'WhitsonClient' generic `get()` and `post()` methods from class -- rather extend functionality from
+  `APIClient`. Added properties for fields, wells, projects, and production data.
+- Added `FieldsAPI` class to `list()` and `retrieve()` fields from a Whitson project
+- Added `ProductionAPI` class to `retrieve()` production data for a specific well from a Whitson project. Added
+  placeholders for future methods.
+- Added `ProjectsAPI` class to `list()` and `retrieve()` projects from a Whitson project
+- Added `WellsAPI` class to `list()` wells from a Whitson project. Also added `run_bhp_calcs()` and
+  `retrieve_bhp_calcs()`
+- Added Python Dataclasses to conform retrieved data to a specified format
+
 ## [0.1.1] - 2023-08-16
 ### Fixed
 - Documentation on PyPI